[Decomp][Spell] Cohesion split: extract shapeshift-boost/absorb/misc aura handlers to SpellAuraMisc.cpp
Move the contiguous tail of the AURA EFFECTS section (HandleShapeshiftBoosts
through HandleSchoolAbsorb, 16 handlers: shapeshift-spell boosts, school
absorb shields, empathy, untrackable, pacify, ghost, allow-flight,
expertise/rating-from-stat, combo-point retain, spirit of redemption) out of
SpellAuras.cpp into a new SpellAuraMisc.cpp translation unit.

Pure file-level cohesion split: declarations stay in SpellAuras.h, the
AuraHandler[] dispatch table stays in SpellAuras.cpp and resolves the moved
member methods by address at link time. Byte-identical move, no behaviour
change. SpellAuras.cpp 12994 -> 12245.
